Clean up: removed old expressions over metamodel
