+++ packages/meta/src/__tests__/expression-tests/checkExpressionChecker.test.ts
@@ -1,15 +1,15 @@
 import { LanguageParser } from "../../languagedef/parser/LanguageParser.js";
 import { FreMetaLanguage } from '../../languagedef/metalanguage';
-import { LanguageExpressionTesterNew } from "../../langexpressions/parser/LanguageExpressionTesterNew.js";
-import { LanguageExpressionParserNew } from "../../langexpressions/parser/LanguageExpressionParserNew.js";
+import { LanguageExpressionTester } from "../../langexpressions/parser/LanguageExpressionTester.js";
+import { LanguageExpressionParser } from "../../langexpressions/parser/LanguageExpressionParser.js";
 import { describe, test, expect, beforeEach } from "vitest";
 import { MetaLogger } from '../../utils/no-dependencies/index.js';
 import { Checker } from '../../utils/basic-dependencies/index.js';
 
 describe("Checking the expression checker", () => {
-    const testdir = "src/__tests__/new-expression-tests/expressionDefFiles/";
-    let parser: LanguageExpressionParserNew;
-    let checker: Checker<LanguageExpressionTesterNew>;
+    const testdir = "src/__tests__/expression-tests/expressionDefFiles/";
+    let parser: LanguageExpressionParser;
+    let checker: Checker<LanguageExpressionTester>;
     let language: FreMetaLanguage | undefined;
     MetaLogger.muteAllErrors();
     MetaLogger.muteAllLogs();
@@ -22,7 +22,7 @@ describe("Checking the expression checker", () => {
                 console.log("Language could not be read");
             }
         }
-        parser = new LanguageExpressionParserNew(language);
+        parser = new LanguageExpressionParser(language);
         checker = parser.checker;
     });
