+++ packages/meta/src/__tests__/expression-tests/expressionDefFiles/demoExpressions.fretest
@@ -12,11 +12,11 @@ DemoEntity {
     self.int_functions;
     self.attrutes;
     self.functions;
-    //#DemoAttributeType:Inter;
-    container;
-    typeof ( self.attributes );
-    typeof ( container );
-    typeof ( xx );
+    #DemoAttributeType:Inter;
+    owner();
+    self.attributes.type (  );
+    owner().type (  );
+    xx.type ( );
 }
 
 DemoAttribute {
@@ -24,7 +24,7 @@ DemoAttribute {
 }
 
 DemoFunction {
-    conformsTo (self.declaredType);
+    self.declaredType.conformsTo ();
     self.parameters;
     DemoFunction.parameters;
 }
@@ -34,46 +34,46 @@ DemoVariable {
 }
 /*
 DemoAbsExpression {
-    equalsType( self.expr, #DemoAttributeType:Integer ) ;
+    self.expr.equalsType( #DemoAttributeType:Integer ) ;
 }
 
 DemoBinaryExpression {
-    equalsType( self.left, #DemoAttributeType:Integer );
-    equalsType( self.right, #DemoAttributeType:Integer );
+    self.left.self.left.equalsType( #DemoAttributeType:Integer );
+    self.right.equalsType( #DemoAttributeType:Integer );
 }
 DemoMultiplyExpression {
-    equalsType( self.left, #DemoAttributeType:Integer );
-    equalsType( self.right, #DemoAttributeType:Integer );
+    self.left.equalsType( #DemoAttributeType:Integer );
+    self.right.equalsType( #DemoAttributeType:Integer );
 }
 
 DemoPlusExpression {
-    equalsType( self.left, #DemoAttributeType:Integer );
-    equalsType( self.right, #DemoAttributeType:Integer );
-    conformsTo( self.left, self.right );
+    self.left.equalsType( #DemoAttributeType:Integer );
+    self.right.equalsType( #DemoAttributeType:Integer );
+    self.left.conformsTo( self.right );
 }
 
 DemoDivideExpression {
-    equalsType( self.left, #DemoAttributeType:Integer );
-    equalsType( self.right, #DemoAttributeType:Integer );
+    self.left.equalsType( #DemoAttributeType:Integer );
+    self.right.equalsType( #DemoAttributeType:Integer );
 }
 
 DemoAndExpression {
-    equalsType( self.left, #DemoAttributeType:Boolean );
-    equalsType( self.right, #DemoAttributeType:Boolean );
+    self.left.equalsType( #DemoAttributeType:Boolean );
+    self.right.equalsType( #DemoAttributeType:Boolean );
 }
 
 DemoOrExpression {
-    equalsType( self.left, #DemoAttributeType:Boolean );
-    equalsType( self.right, #DemoAttributeType:Boolean );
+    self.left.equalsType( #DemoAttributeType:Boolean );
+    self.right.equalsType( #DemoAttributeType:Boolean );
 }
 
 DemoComparisonExpression {
-    equalsType( self.left, self.right );
+    self.left.equalsType( self.right );
 }
 
 DemoIfExpression {
-    equalsType( self.condition, #DemoAttributeType:Boolean );
-    conformsTo( self.whenTrue, self.whenFalse );
+    self.condition.equalsType( #DemoAttributeType:Boolean );
+    self.whenTrue.conformsTo( self.whenFalse );
 }
 
 */
