--- packages/meta/src/__tests__/new-expression-tests/expressionDefFiles/demoExpressions.fretest
@@ -1,79 +1,0 @@
-expressions for language Demo
-
-DemoModel {
-    // default Identifier
-    self.entities.expr;
-    self.functions;
-}
-
-DemoEntity {
-    self.simpleprop;
-    self.int_attr;
-    self.int_functions;
-    self.attrutes;
-    self.functions;
-    #DemoAttributeType:Inter;
-    owner();
-    self.attributes.type (  );
-    owner().type (  );
-    xx.type ( );
-}
-
-DemoAttribute {
-     DemoVariable.fgg;
-}
-
-DemoFunction {
-    self.declaredType.conformsTo ();
-    self.parameters;
-    DemoFunction.parameters;
-}
-
-DemoVariable {
-    self.extra;
-}
-/*
-DemoAbsExpression {
-    self.expr.equalsType( #DemoAttributeType:Integer ) ;
-}
-
-DemoBinaryExpression {
-    self.left.self.left.equalsType( #DemoAttributeType:Integer );
-    self.right.equalsType( #DemoAttributeType:Integer );
-}
-DemoMultiplyExpression {
-    self.left.equalsType( #DemoAttributeType:Integer );
-    self.right.equalsType( #DemoAttributeType:Integer );
-}
-
-DemoPlusExpression {
-    self.left.equalsType( #DemoAttributeType:Integer );
-    self.right.equalsType( #DemoAttributeType:Integer );
-    self.left.conformsTo( self.right );
-}
-
-DemoDivideExpression {
-    self.left.equalsType( #DemoAttributeType:Integer );
-    self.right.equalsType( #DemoAttributeType:Integer );
-}
-
-DemoAndExpression {
-    self.left.equalsType( #DemoAttributeType:Boolean );
-    self.right.equalsType( #DemoAttributeType:Boolean );
-}
-
-DemoOrExpression {
-    self.left.equalsType( #DemoAttributeType:Boolean );
-    self.right.equalsType( #DemoAttributeType:Boolean );
-}
-
-DemoComparisonExpression {
-    self.left.equalsType( self.right );
-}
-
-DemoIfExpression {
-    self.condition.equalsType( #DemoAttributeType:Boolean );
-    self.whenTrue.conformsTo( self.whenFalse );
-}
-
-*/
