+++ packages/meta/src/__tests__/typer-tests/checkPlaygroundDefs.test.ts
@@ -6,7 +6,7 @@ import { FreTyperMerger } from '../../typerdef/parser';
 import { describe, test, expect, beforeEach } from "vitest";
 import { FretClassifierSpec } from "../../typerdef/metalanguage"
 
-describe("Checking new typer", () => {
+describe("Checking typer", () => {
     const testdir = "src/__tests__/typer-tests/";
     let parser: FreTyperMerger;
     let language: FreMetaLanguage | undefined;
