+++ packages/meta/src/editordef/parser/concat.sh
@@ -2,4 +2,4 @@
 
 LANGDEV=../../languagedef/parser
 
-cat  FreEditGrammar.part.pegjs $LANGDEV/ExpressionGrammar.part.pegjs $LANGDEV/BasicGrammar.part.pegjs > FreEditGrammar.pegjs
+cat  FreEditGrammar.part.pegjs $LANGDEV/BasicGrammar.part.pegjs > FreEditGrammar.pegjs
