+++ packages/meta/src/langexpressions/metalanguage/FreLangExpressions.ts
@@ -19,7 +19,7 @@ import { MetaFunctionNames } from '../../utils/no-dependencies/index.js';
 // These properties need to be undefined during parsing and checking. After the checking process
 // has been executed without errors, we can assume that these properties are initialized.
 
-export abstract class FreLangExpNew extends FreMetaLangElement {
+export abstract class FreLangExp extends FreMetaLangElement {
     // @ts-ignore
     language: FreMetaLanguage; // the language for which this expression is defined
 
@@ -27,7 +27,7 @@ export abstract class FreLangExpNew extends FreMetaLangElement {
      * Returns a string that is equal to the input of the parser for this expression.
      */
     toFreString(): string {
-        return 'SHOULD BE IMPLEMENTED BY SUBCLASSES OF "FreLangExpressionsNew.FreLangExpNew"';
+        return 'SHOULD BE IMPLEMENTED BY SUBCLASSES OF "FreLangExpressions.FreLangExp"';
     }
 
     /**
@@ -66,13 +66,13 @@ export abstract class FreLangExpNew extends FreMetaLangElement {
     /**
      * Returns the last of the applied expressions.
      */
-    getLastExpression(): FreLangExpNew {
+    getLastExpression(): FreLangExp {
         // should be overridden by all subclasses that can have an applied expression
         return this;
     }
 }
 
-export abstract class FreVarOrFunctionExp extends FreLangExpNew {
+export abstract class FreVarOrFunctionExp extends FreLangExp {
     // @ts-ignore
     name: string;
     applied: FreVarOrFunctionExp | undefined;
@@ -111,7 +111,7 @@ export abstract class FreVarOrFunctionExp extends FreLangExpNew {
         return undefined;
     }
 
-    getLastExpression(): FreLangExpNew {
+    getLastExpression(): FreLangExp {
         if (this.applied) {
             return this.applied.getLastExpression();
         } else {
@@ -173,7 +173,7 @@ export class FreVarExp extends FreVarOrFunctionExp {
  */
 export class FreFunctionExp extends FreVarOrFunctionExp {
     // @ts-ignore
-    param: FreLangExpNew | undefined;
+    param: FreLangExp | undefined;
     possibleClassifiers: FreMetaClassifier[] = [];
 
     toFreString(): string {
@@ -200,7 +200,7 @@ export class FreFunctionExp extends FreVarOrFunctionExp {
 /**
  * This class represents an expression that refers to an instance of a limited concept.
  */
-export class FreLimitedInstanceExp extends FreLangExpNew {
+export class FreLimitedInstanceExp extends FreLangExp {
     conceptName: string = ''; // should be the name of a limited concept
     instanceName: string = ''; // should be the name of one of the predefined instances of 'sourceName'
     // @ts-ignore
@@ -238,7 +238,7 @@ export class FreLimitedInstanceExp extends FreLangExpNew {
 /**
  * This class represents a simple expression, currently just a number.
  */
-export class FreLangSimpleExpNew extends FreLangExpNew {
+export class FreLangSimpleExp extends FreLangExp {
     // @ts-ignore
     value: number;
