+++ packages/meta/src/langexpressions/metalanguage/index.ts
@@ -1,1 +1,1 @@
-export * from "./FreLangExpressionsNew.js";
+export * from "./FreLangExpressions.js";
