+++ packages/meta/src/languagedef/metalanguage/internal.ts
@@ -6,5 +6,4 @@
 
 export * from "./MetaElementReference.js";
 export * from "./FreMetaLanguage.js";
-export * from "./FreLangExpressions.js";
 export * from "./LangUtil.js";
