+++ packages/meta/src/languagedef/parser/concat.sh
@@ -1,4 +1,3 @@
 #!/usr/bin/env bash
 
 cat  LanguageGrammar.part.pegjs BasicGrammar.part.pegjs > LanguageGrammar.pegjs
-cat  ExpressionGrammar.part.header.pegjs ExpressionGrammar.part.pegjs BasicGrammar.part.pegjs > ExpressionGrammar.pegjs
