+++ packages/meta/src/scoperdef/metalanguage/FreScopeDefLang.ts
@@ -1,6 +1,6 @@
 import { FreMetaClassifier, MetaElementReference } from '../../languagedef/metalanguage/index.js';
 import { FreMetaDefinitionElement } from "../../utils/no-dependencies/index.js";
-import { FreLangExpNew } from '../../langexpressions/metalanguage/index.js';
+import { FreLangExp } from '../../langexpressions/metalanguage/index.js';
 import { isNullOrUndefined } from '../../utils/file-utils/index.js';
 
 export class ScopeDef extends FreMetaDefinitionElement {
@@ -77,7 +77,7 @@ export class FreNamespaceAlternative extends FreMetaDefinitionElement {
 }
 
 export class FreMetaNamespaceInfo extends FreMetaDefinitionElement {
-    expression: FreLangExpNew | undefined;
+    expression: FreLangExp | undefined;
     recursive: boolean = false;
 
     toFreString(): string {
